refactor: whole buildchain, drop unneeded deps and scripts
