--- config/webpack.config.client-development.js
@@ -1,8 +1,0 @@
-// @ts-check
-const webpackBaseConfig = require('./webpack.config.base');
-
-module.exports = Object.assign({}, webpackBaseConfig, {
-  target: "web",
-  mode: "development",
-  devtool: "inline-source-map",
-});
