+++ package.json
@@ -12,7 +12,7 @@
     "node",
     "sdk"
   ],
-  "homepage": "https://docs.directus.io/sdk/js.html",
+  "homepage": "https://docs.directus-sdk.io/sdk/js.html",
   "bugs": {
     "url": "https://github.com/directus/directus-sdk-js/issues"
   },
@@ -26,21 +26,25 @@
   "files": [
     "dist"
   ],
-  "main": "dist/directus-sdk.js",
-  "browser": "dist/directus-sdk.umd.min.js",
+  "main": "dist/cjs/index.js",
+  "module": "dist/esm/index.js",
+  "browser": "dist/umd/directus-sdk.umd.min.js",
   "types": "dist/types/index.d.ts",
   "scripts": {
-    "prebuild": "npm-run-all prebuild:*",
-    "build": "run-p build:*",
-    "build:declaration": "tsc --project ./tsconfig.meta.json",
-    "build:webpack": "webpack",
-    "develop": "webpack --progress --colors --watch",
+    "prebuild": "rimraf dist",
+    "build": "run-s build:*",
+    "build:declaration": "tsc --project tsconfig.meta.json",
+    "build:es2015": "tsc --project tsconfig.json --module es2015 --target es2015 --outDir dist/es2015",
+    "build:esm": "tsc --project tsconfig.json --module es2015 --target es5 --outDir dist/esm",
+    "build:cjs": "tsc --project tsconfig.json --module commonjs --target es5 --outDir dist/cjs",
+    "build:umd": "rollup dist/esm/index.js --format umd --name Directus --sourcemap --file dist/umd/directus-sdk.js --globals window --exports named --external axios,base-64,qs/lib/stringify",
+    "build:umd-min": "uglifyjs --output dist/umd/directus-sdk.min.js --mangle --compress --source-map --source-map-url -- dist/umd/directus-sdk.js",
+    "postbuild": "size-limit",
+    "size-stats": "size-limit --why",
     "documentation": "typedoc --theme markdown --out docs --mdEngine github --readme none --sourcefile-url-prefix \"https://github.com/direcuts/sdk-js/tree/master/\"",
     "lint": "run-s lint:*",
     "lint:source": "eslint ./src/**/*.ts -c ./.eslintrc.js",
     "lint:test": "eslint ./test/**/*.ts -c ./.eslintrc.test.js",
-    "prebuild:clean": "rimraf dist",
-    "prebuild:lint": "run-s lint test",
     "prettier": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\" \"**/*.json\"",
     "test": "ts-mocha -p ./tsconfig.test.json \"test/**/*.spec.ts\" --recursive --exit",
     "coverage": "nyc mocha -r ts-node/register -r source-map-support/register  -p ./tsconfig.test.json \"test/**/*.spec.ts\" --recursive && nyc report"
@@ -73,8 +77,10 @@
     "nyc": "14.1.1",
     "prettier": "1.18.2",
     "rimraf": "2.6.3",
+    "rollup": "1.17.0",
     "sinon": "7.3.2",
     "sinon-chai": "3.3.0",
+    "size-limit": "1.3.7",
     "source-map-support": "0.5.12",
     "ts-loader": "6.0.4",
     "ts-mocha": "6.0.0",
@@ -85,32 +91,27 @@
     "typedoc-plugin-markdown": "2.0.8",
     "typedoc-plugin-sourcefile-url": "1.0.4",
     "typescript": "3.5.3",
-    "webpack": "4.35.3",
-    "webpack-bundle-analyzer": "3.3.2",
-    "webpack-cli": "3.3.5",
-    "webpack-node-externals": "1.7.2"
+    "uglify-js": "3.6.0"
   },
   "engines": {
     "node": ">=8.0.0 <13.0.0"
   },
-  "nyc": {
-    "extension": [
-      ".ts"
-    ],
-    "reporter": [
-      "lcov",
-      "text-summary"
-    ],
-    "exclude": [
-      "node_modules",
-      "docs",
-      "dist",
-      "coverage",
-      "config",
-      "test",
-      "**/*.d.ts",
-      "*.spec.ts"
-    ],
-    "all": true
-  }
+  "size-limit": [
+    {
+      "limit": "12 KB",
+      "path": "dist/umd/directus-sdk.min.js"
+    },
+    {
+      "limit": "13 KB",
+      "path": "dist/cjs/index.js"
+    },
+    {
+      "limit": "12 KB",
+      "path": "dist/es2015/index.js"
+    },
+    {
+      "limit": "12 KB",
+      "path": "dist/esm/index.js"
+    }
+  ]
 }
