+++ tsconfig.json
@@ -6,9 +6,15 @@
     "moduleResolution": "node",
     "outDir": "dist",
     "sourceMap": true,
-    "lib": ["dom", "es2015"],
-    "rootDir": ".",
-    "baseUrl": "./src"
+    "lib": [
+      "dom",
+      "es2015"
+    ],
+    "rootDir": "src",
+    "baseUrl": "."
   },
-  "exclude": ["test", "dist"]
+  "exclude": [
+    "test",
+    "dist"
+  ]
 }
